+++ LuxEngine/Core/Memory/VRam/VClasses.hpp
@@ -41,7 +41,7 @@ namespace lux::vram{
 				#define _case(n) case VCellClass::CLASS_##n: return (uint16)VCellClassIndex::INDEX_##n;
 				case VCellClass::CLASS_0: return (uint16)-1;			//Custom size class
 				_case(A) _case(B) _case(C) _case(D) _case(Q) _case(L)	//Fixed size classes
-				default: return -2; 									//This is just to suppress the warning
+				default: return (uint16)-2; 							//This is just to suppress the warning
 				#undef _case
 			}
 		}
